Demo of training ConvAE on landcover map implemented with new data module logic
