+++ configs/config_convae.yaml
@@ -1,0 +1,19 @@
+# configs/config_convae.yaml
+defaults:
+  - data: demonstrator_datasets_landcover
+  - model: demo_convae
+  - trainer: default
+  - logging: mlflow_convae_demo
+  - _self_
+
+seed: 42
+
+dataset_chunks:   ${oc.env:FRAME_DATASET_CHUNKS, "auto"}
+caching_backend:  ${oc.env:FRAME_CACHING_BACKEND, "basic"}      # Or "series", "dask_distributed" or "slurm" - which will use the `zarr-parallel` library.
+zarr_format:      ${oc.env:FRAME_ZARR_FORMAT, 2}
+
+hydra:
+  run:
+    dir: outputs/${now:%Y-%m-%d}/${now:%H-%M-%S}
+  job:
+    chdir: true
