+++ configs/model/demo_convae.yaml
@@ -1,11 +1,13 @@
-# configs/model/ConvAE.yaml
-_target_: FRAME_FM.models.demo_convAE.ConvAutoencoder
+# configs/model/demo_convae.yaml
+_target_: FRAME_FM.models.convae.ConvAutoencoder
 
 # Architecture
 in_channels: 10         # currently set up only for 10-band land cover map
 base_channels: 32       # maps to chs[0] in the first layer of the encoder
 kernel_size: 3          # kernel size for all convolutional layers; currently set to 3 to maintain spatial dimensions across layers (with padding=1)
-latent_dim: 1024         # dimension of the latent space
+latent_dim: 514         # dimension of the latent space
+input_dim: 32           # spatial dimension of the input (assumed to be square, i.e. 32x32)
+plotting: True          # whether to plot input and reconstructed occasionally
 
 
 # Optimisation hyperparameters
