--- configs/trainer/convAE_trainer.yaml
@@ -1,17 +1,0 @@
-# configs/trainer/convAE_trainer.yaml
-_target_: pytorch_lightning.Trainer
-
-max_epochs: 1000
-accelerator: auto      # "auto", "gpu", "cpu"
-devices: auto          # or e.g. 1, [0], etc.
-precision: 32          # or "16-mixed" for AMP
-
-log_every_n_steps: 5
-enable_checkpointing: true
-enable_progress_bar: true
-enable_model_summary: true
-
-# Optional extra flags:
-# gradient_clip_val: 0.0
-# deterministic: true
-# run_test: true  # flag for evaluating on test set after training
