+++ scripts/landcover_convae_demo
@@ -1,0 +1,17 @@
+#!/bin/bash
+#SBATCH --job-name="ConvAE train"
+#SBATCH --time=10:00:00
+#SBATCH --mem=16G
+#SBATCH --partition=orchid
+#SBATCH --account=orchid
+#SBATCH --qos=orchid
+#SBATCH --gres=gpu:1
+#SBATCH -o %j.out
+#SBATCH -e %j.err
+
+# executable
+[ ! "$VIRTUAL_ENV" ] &&  "source .venv/bin/activate"
+export HYDRA_FULL_ERROR=1
+export PYTHONPATH="$PWD/src:$PYTHONPATH" # this is necessary at the moment because baseline modules are in src/
+# python src/FRAME_FM/training/train.py model=demo_mmmae data=spatial_land_cover trainer=default logging=mlflow_mmmae_demo
+python src/FRAME_FM/training/train.py model=demo_convae data=demonstrator_datasets_landcover trainer=default logging=mlflow_convae_demo
