+++ src/FRAME_FM/training/logging_utils.py
@@ -1,6 +1,6 @@
 # src/FRAME_FM/training/logger.py
 
-from pytorch_lightning.loggers import MLFlowLogger
+from pytorch_lightning.loggers import CSVLogger, MLFlowLogger 
 from typing import Any, Optional
 
 def create_mlflow_logger(
